clean up remnants of moving functions
